# api/openapi.yaml
openapi: 3.0.3
info:
  title: Mayastor RESTful API
  version: v0
servers:
- url: /v0
paths:
...
  /nodes/{id}/replicas:
    get:
      operationId: get_node_replicas
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Replica'
                type: array
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Request Timeout
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Found
        "408":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Bad Request
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Precondition Failed
        "416":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Range Not satisfiable
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unprocessable entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Implemented
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Service Unavailable
        "504":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Gateway Timeout
        "507":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Insufficient Storage
      security:
      - JWT: []
      tags:
      - Replicas
...
  /nodes/{node_id}/pools/{pool_id}/replicas:
    get:
      operationId: get_node_pool_replicas
      parameters:
      - explode: false
        in: path
        name: node_id
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: pool_id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Replica'
                type: array
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Request Timeout
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Found
        "408":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Bad Request
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Precondition Failed
        "416":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Range Not satisfiable
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unprocessable entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Implemented
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Service Unavailable
        "504":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Gateway Timeout
        "507":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Insufficient Storage
      security:
      - JWT: []
      tags:
      - Replicas
  /nodes/{node_id}/pools/{pool_id}/replicas/{replica_id}:
    delete:
      operationId: del_node_pool_replica
      parameters:
      - explode: false
        in: path
        name: node_id
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: pool_id
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: replica_id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "204":
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Request Timeout
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Found
        "408":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Bad Request
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Precondition Failed
        "416":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Range Not satisfiable
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unprocessable entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Implemented
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Service Unavailable
        "504":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Gateway Timeout
        "507":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Insufficient Storage
      security:
      - JWT: []
      tags:
      - Replicas
    get:
      operationId: get_node_pool_replica
      parameters:
      - explode: false
        in: path
        name: node_id
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: pool_id
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: replica_id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Replica'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Request Timeout
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Found
        "408":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Bad Request
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Precondition Failed
        "416":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Range Not satisfiable
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unprocessable entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Implemented
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Service Unavailable
        "504":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Gateway Timeout
        "507":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Insufficient Storage
      security:
      - JWT: []
      tags:
      - Replicas
    put:
      operationId: put_node_pool_replica
      parameters:
      - explode: false
        in: path
        name: node_id
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: pool_id
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: replica_id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReplicaBody'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Replica'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Request Timeout
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Found
        "408":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Bad Request
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Precondition Failed
        "416":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Range Not satisfiable
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unprocessable entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Implemented
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Service Unavailable
        "504":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Gateway Timeout
        "507":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Insufficient Storage
      security:
      - JWT: []
      tags:
      - Replicas
  /nodes/{node_id}/pools/{pool_id}/replicas/{replica_id}/share:
    delete:
      operationId: del_node_pool_replica_shares
      parameters:
      - explode: false
        in: path
        name: node_id
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: pool_id
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: replica_id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "204":
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Request Timeout
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Found
        "408":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Bad Request
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Precondition Failed
        "416":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Range Not satisfiable
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unprocessable entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Implemented
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Service Unavailable
        "504":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Gateway Timeout
        "507":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Insufficient Storage
      security:
      - JWT: []
      tags:
      - Replicas
  /nodes/{node_id}/pools/{pool_id}/replicas/{replica_id}/share/{protocol}:
    put:
      operationId: put_node_pool_replica_share
      parameters:
      - explode: false
        in: path
        name: node_id
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: pool_id
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: replica_id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      - explode: false
        in: path
        name: protocol
        required: true
        schema:
          enum:
          - nvmf
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                type: string
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Request Timeout
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Found
        "408":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Bad Request
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Precondition Failed
        "416":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Range Not satisfiable
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unprocessable entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Implemented
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Service Unavailable
        "504":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Gateway Timeout
        "507":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Insufficient Storage
      security:
      - JWT: []
      tags:
      - Replicas
...
  /pools/{pool_id}/replicas/{replica_id}:
    delete:
      operationId: del_pool_replica
      parameters:
      - explode: false
        in: path
        name: pool_id
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: replica_id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "204":
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Request Timeout
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Found
        "408":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Bad Request
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Precondition Failed
        "416":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Range Not satisfiable
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unprocessable entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Implemented
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Service Unavailable
        "504":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Gateway Timeout
        "507":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Insufficient Storage
      security:
      - JWT: []
      tags:
      - Replicas
    put:
      operationId: put_pool_replica
      parameters:
      - explode: false
        in: path
        name: pool_id
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: replica_id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReplicaBody'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Replica'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Request Timeout
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Found
        "408":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Bad Request
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Precondition Failed
        "416":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Range Not satisfiable
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unprocessable entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Implemented
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Service Unavailable
        "504":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Gateway Timeout
        "507":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Insufficient Storage
      security:
      - JWT: []
      tags:
      - Replicas
  /pools/{pool_id}/replicas/{replica_id}/share:
    delete:
      operationId: del_pool_replica_shares
      parameters:
      - explode: false
        in: path
        name: pool_id
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: replica_id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "204":
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Request Timeout
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Found
        "408":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Bad Request
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Precondition Failed
        "416":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Range Not satisfiable
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Unprocessable entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Internal Server Error
        "501":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Not Implemented
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Service Unavailable
        "504":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Gateway Timeout
        "507":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestJsonError'
          description: Insufficient Storage
      security:
      - JWT: []
      tags:
      - Replicas
...
components:
  schemas:
...
    CreateReplicaBody:
      description: Create Replica Body JSON
      example:
        share: "off"
        size: 0
        thin: false
      properties:
        share:
          description: protocol to expose the replica over
          enum:
          - "false"
          - nvmf
          - iscsi
          - nbd
          type: string
        size:
          description: size of the replica in bytes
          format: int64
          type: integer
        thin:
          description: thin provisioning
          type: boolean
      required:
      - share
      - size
      - thin
      type: object
...
    Replica:
      description: Replica information
      example:
        node: ""
        pool: ""
        share: "off"
        size: 0
        state: unknown
        thin: false
        uri: ""
        uuid: 514ed1c8-7174-49ac-b9cd-ad44ef670a67
      properties:
        node:
          description: id of the mayastor instance
          type: string
        pool:
          description: id of the pool
          type: string
        share:
          description: protocol used for exposing the replica
          enum:
          - "false"
          - nvmf
          - iscsi
          - nbd
          type: string
        size:
          description: size of the replica in bytes
          format: int64
          type: integer
        state:
          description: state of the replica
          enum:
          - unknown
          - online
          - degraded
          - faulted
          type: string
        thin:
          description: thin provisioning
          type: boolean
        uri:
          description: uri usable by nexus to access it
          type: string
        uuid:
          description: uuid of the replica
          format: uuid
          type: string
      required:
      - node
      - pool
      - share
      - size
      - state
      - thin
      - uri
      - uuid
      type: object
    RestJsonError:
      description: Rest Json Error format
      example:
        details: ""
        error: NotFound
      properties:
        details:
          description: detailed error information
          type: string
        error:
          description: error kind
          enum:
          - Timeout
          - Deserialize
          - Internal
          - InvalidArgument
          - DeadlineExceeded
          - NotFound
          - AlreadyExists
          - PermissionDenied
          - ResourceExhausted
          - FailedPrecondition
          - NotShared
          - NotPublished
          - AlreadyPublished
          - AlreadyShared
          - Aborted
          - OutOfRange
          - Unimplemented
          - Unavailable
          - Unauthenticated
          - Unauthorized
          - Conflict
          - FailedPersist
          - Deleting
          type: string
      required:
      - details
      - error
      type: object
...
  securitySchemes:
    JWT:
      bearerFormat: JWT
      scheme: bearer
      type: http